Added silent args, group command, and moved listeners
- All listeners that aren't message have been moved to a separate file. In future may make them it a folder, and each listener type have it's own file. Not yet decided
-silent args group is mainly to be used to pass arguments when the command is being used without invoking the prefix such as emotes
- group command is basically my solution to handle duel commands without using the a hyphen between them. May be used to refactor future commands

+++ src/commands/tournament/signup.ts
@@ -12,7 +12,7 @@ export const signup: Command = {
     async execute(message: Message, client: Client, args: string[]) {
         let signup:Signups = await getDoc("config", "signups")
 
-        if(message.channel.type !== "dm" && message.id !== signup.msgID) {
+        if(message.channel.type !== "dm" && message.id !== signup.msgID && message.author.id === client.user?.id) {
             return await message.reply("You have to signup in bot dm if they are open").then(async m =>{
                 message.delete()
                 m.delete({timeout:1600})
